Supporting CaseInsensitive Queries
We have a DNS query `test.cloud.example.com` which returns a list of IP addresses. What should happen when you
query `test.cloud.EXAMPLE.COM`. The answer is in the RFC https://www.rfc-editor.org/rfc/rfc4343#section-4.2.
We return the same list of IP addresses for both queries. This also implies that if we maintain two dnsendpoints
`test.cloud.example.com` with address `1.1.1.1` and `test.cloud.EXAMPLE.COM` with address `2.2.2.2`, we must return
the merge `[1.1.1.1 2.2.2.2]` for both queries.

Signed-off-by: Michal Kuritka <kuritka@gmail.com>