Configure renovate bot (#83)
Let's give coredns-crd-plugin some love and configure Renovate bot, as we did in https://github.com/k8gb-io/k8gb. Both repositories will be used the same dependency management tool, which will be easier to maintain.

The configuration is similar to the k8gb repository. We group all non major upgrades in one PR, create a dedicated PR for other bumps, and pin digests of Github Actions.

In addition, we prevent the upgrade of caddy to v2 since that version is not maintained, as described in 855032f.

We can also disable dependabot, so we remove its configuration.

Signed-off-by: Andre Aguas <andre.aguas@protonmail.com>